pyproject.toml
[project]
name = "burnbag"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "boto3",
    "psycopg2-binary",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]